requirements.txt
langchain==0.3.17
pydantic==2.10.6
langchain-openai==0.3.2
PyGithub